added @Testable to document binding in order to enable IDE support
The dependency to the JUnit Platform is compile time only and therefore not transitive. No one HAS to have JUnit on the class path in order to use @ExecutableDocument, but if our JUnit Engine is used, classes annotated with it will be recognized by the IDE as "testable" and allow an execution.