Enhance step alignment with variable quotation
This makes the alignment algorithm aware of optional quotation characters for
variable separation. They can be used for disambiguation in case of an
insertion that is adjacent to, but not supposed to be part of a variable.

Given the following alignment:
```
    Peter likes {food}.
    Peter likes -Pizza. (food = "Pizza")
```

Suppose, we insert an adjective in front of food:
```
    Peter likes {     food    }.
    Peter likes delicious Pizza. (food = "delicious Pizza")
```

If we don't want the insertion included, we can use quotation:
```
    Peter likes --------- '{food}'.
    Peter likes delicious '-Pizza'. (food = "Pizza")
```