Prevent quadratic time lookups
The previous implementation iterated over all rows results to find one
that matches a row of the decision table. This lookup was done once for
each row of the decision table, resulting in quadratic time
requirements.

The way our results are constructed, however, guarantees that row
results will be added into the builder in iteration order of the rows of
the decision table. The new implementation takes advantage of this by
only comparing values at the same position in the decision table and
the list of results.